Clean framer/__init__.py
